==> 1. titanic_eda.ipynb <==
# Understanding Data

### 4 Things to Focus on:-
    1:- Basic Question to Ask (Preprocessing)
    2:- EDA (Univariate, Bivariate)
    3:- EDA (Multi Variate)
    4:- Pandas Profiler (To Automote  these task)

import pandas as pd

df = pd.read_csv('titanic.csv')

# Preprocessing Steps

### 1. How big is the Data ?

df.shape

### 2. How does the data look like ?

# df.head()

df.sample(5)

### 3. What is the data type of cols ?

df. info()

### 4. Are there any missing value ?

df.isnull().sum()

###  5. How does the data look mathematically ?

df.describe()

### 6. Are there duplicate values ?

df.duplicated().sum()

### 7. How is the correlation between cols ?

df.corr(numeric_only=True)

# Checking correlation of Survived  Column with  other columns

df.corr(numeric_only=True)['Survived']

# Exploratory Data Analysis

### Best ways to data EDA are:-
    1:- Univariate Analysis
    2:- Bivariate Analysis
    3:- Multivariate Analyis
