==> 4. Pandas_Profiling EDA.ipynb <==
import pandas as pd

df = pd.read_csv('titanic.csv')

df.head()

from ydata_profiling import ProfileReport

profile = ProfileReport(df, title="Profiling Report")

profile.to_file('Report.html')


# prof = ProfileReport(df)
# prof.to_file(output_file='output.html')
