==> 8. Ordinal_Encoding.ipynb <==
# Ordinal and Lable Encoding

import numpy as np
import pandas as pd

df = pd.read_csv('customer.csv')

df.sample(5)

df = df.iloc[:,2:]

df.head()

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(df.iloc[:,0:2], df.iloc[:,-1], test_size=0.2)


X_train

from sklearn.preprocessing import OrdinalEncoder

oe = OrdinalEncoder(categories=[['Poor','Average', 'Good'], ['School','UG','PG']])

oe.fit(X_train)

X_train = oe.transform(X_train)
X_test = oe.transform(X_test)

X_train

oe.categories_

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

le.fit(y_train)

le.classes_

y_train = le.transform(y_train)
y_test = le.transform(y_test)

y_train
