==> 9. OneHotEncoding.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('cars.csv')

df.head()

# Total values counts in each brands

df['brand'].value_counts()

# Total Unique brand counts

df['brand'].nunique()

# Total values counts in each fue

df['fuel'].value_counts()

# Total values counts in each fue

df['fuel'].value_counts()

# Total values counts in each owner

df['owner'].value_counts()

# 1. OneHotEncoding using Pandas

pd.get_dummies(df, columns=['fuel', 'owner'], dtype=int)

# 2. K-1 OneHotEncoding

pd.get_dummies(df, columns=['fuel', 'owner'], dtype=int, drop_first=True)

### Note:- At the time of Data analysis we can use pandas and get_dummies(). But we can not use this at the time of machine learning project. The reason for not using the get_dummies() in machine learning project is that because pandas does not remember which column is at which position. When we ran the  code with pandas.get_dummies() result could  be change each and every time. So that's why we use "Sklearn OneHotEncoder".

# 3. OneHotEncoding using Sklearn

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(df.iloc[:,0:4], df.iloc[:,-1], test_size=0.2, random_state=0)

X_train.head()

from sklearn.preprocessing import OneHotEncoder

ohe = OneHotEncoder(drop='first', sparse=False, dtype=np.int32)

#### Note:- Now we have to take out 'fuel' and 'owner' column from dataset and perform OneHotEncoding after appling OneHotEncoding then we have to add those encoded column into dataset.
#### To do this step in short way we have to use column Transformer but in this session we do not use Column Transformer we will do it manually. We will see  column Transformer  later. 

# X_train_new = ohe.fit_transform(X_train[['fuel','owner']]).toarray()

X_train_new = ohe.fit_transform(X_train[['fuel','owner']])

#### Note:- OneHotEncoder by default  produce sparse matrix to  view the matrix, Either we have to use (toarray()) method or we can use 'Sparse = False ' in 'OneHotEncoder(drop='first', Sparse=False)' then we don't need to use 'toarray()' method. It will automatically give numpy array after using  the 'Sparse = False' parameter.

# X_test_new = ohe.transform(X_test[['fuel','owner']]).toarray()

X_test_new = ohe.transform(X_test[['fuel','owner']])

X_train_new.shape


X_train[['brand','km_driven']].values

### To merge the column we  have to use below code:-

np.hstack((X_train[['brand','km_driven']].values, X_train_new)).shape

#### Note:- To remove the first column from encoded column to save us from multi colinearity we have to use 'OneHotEncoder(drop=True)'.

# 4. OneHotEncoding with Top Categories

counts = df['brand'].value_counts()

df['brand'].nunique()
threshold = 100 

repl = counts[counts <= threshold].index

pd.get_dummies(df['brand'].replace(repl, 'Uncommon'), dtype=int)
