==> 15. Date_Time.ipynb <==
# Working with Date and Time

import numpy as np
import pandas as pd

date = pd.read_csv('orders.csv')
time = pd.read_csv('messages.csv')

date.head()

time.head()

date.info()

### Note:- When we are importing dataset in pandas by default date and time column datatype is object (String). So If it is Object type column then we can not perform Date and Type operation on this column. Our First Priority for such columns is we need to change such column into proper datetime format.

time.info()

## Working with Date

# Converting to datetime datatype

date['date'] = pd.to_datetime(date['date'])

date.info()

#### 1. Extract Year

date['date_year'] = date['date'].dt.year

date.head()

#### 2. Extract Month

date['date_month_no'] = date['date'].dt.month

date.head()

# For Extracting month name

date['date_month_name'] = date['date'].dt.month_name()

date.head()

#### 3. Extracting Days

date['date_day'] = date['date'].dt.day

date.head()

# Extracting  day of week

date['date_DOW'] = date['date'].dt.day_of_week

date.head()

# Extracting Day of Week - Name

date['date_DOW_name'] = date['date'].dt.day_name()

date.drop(columns=['product_id','city_id','orders']).head()

# Is Weekend ?

date['date_is_weekend'] = np.where(date['date_DOW_name'].isin(['Sunday','Saturday']),1,0)

date.drop(columns=['product_id','city_id','orders']).head()

#### 4. Extracting Week of the Year

# date['date_week'] = date['date'].dt.weekday # This works in earlier pandas version

date['date_week'] = date['date'].dt.isocalendar().week

date.drop(columns=['product_id','city_id','orders']).head()

#### 5. Extracting Quarter

date['quarter'] = date['date'].dt.quarter

date.drop(columns=['product_id','city_id','orders']).head()

#### 6. Extract Semester

date['semester'] = np.where(date['quarter'].isin([1,2]),1,2)

date.drop(columns=['product_id','city_id','orders']).head()

#### 7. Extracting Time Elapsed Between Date (Gap(Difference) between Two Dates)

import datetime

today = datetime.datetime.today()

today

today - date['date']

# Extracting Elapsed Days

(today - date['date']).dt.days

# Month Passed

np.round((today - date['date'])/np.timedelta64(1,'M'),0)

## Working With Time

time

time.info()

# Converting to datetime datatype

time['date'] = pd.to_datetime(time['date'])

time.info()

time['hour'] = time['date'].dt.hour
time['min'] = time['date'].dt.minute
time['sec'] = time['date'].dt.second

time.head()

#### 1. Extracting Time Part

time['time'] = time['date'].dt.time

time.head()

#### 2. Extracting Time Difference

today - time['date']

# Time Difference in Seconds

(today - time['date'])/np.timedelta64(1, 's')

# Time Difference in Minutes

(today - time['date'])/np.timedelta64(1, 'm')

# Time Difference in Hours

(today - time['date'])/np.timedelta64(1, 'h')
